+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+ * PyTryFrom's error is always to `PyDowncastError`
+
+### Removed
+
+ * The pyobject_downcast macro
+
 ## [0.4.0] - 2018-07-30
 
 ### Removed